chore: bump to bbn v0.11

+++ itest/test_manager.go
@@ -458,7 +458,7 @@ func (tm *TestManager) InsertCovenantSigForDelegation(t *testing.T, btcDel *bsty
 		[]*btcec.PublicKey{btcDel.FpBtcPkList[0].MustToBTCPK()},
 		params.CovenantPks,
 		params.CovenantQuorum,
-		btcDel.GetStakingTime(),
+		uint16(btcDel.GetStakingTime()),
 		btcutil.Amount(btcDel.TotalSat),
 		simnetParams,
 	)
